Notifications | Divers Correctifs

+++ admin.py
@@ -11,7 +11,7 @@ from auth import get_current_user
 from datetime import date
 from log import api_log
 
-def admin_required(current_user: Annotated[models.Users, Depends(get_current_user)]):
+def admin_required(current_user: Annotated[models.Users, Depends(get_current_user)], request: Request):
     approvedList = ["admin", "owner"]
     if not getattr(current_user, "privileges", "player") in approvedList:
         api_log("admin.access_denied", level="INFO", request=request, tags=["auth", "admin"], user_id=current_user.id,email=current_user.email, correlation_id=request.headers.get("x-correlation-id")) # type: ignore
@@ -26,21 +26,26 @@ router = APIRouter(
 
 bcrypt_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
+dicoAllowToChange = {
+    'owner': ['owner', 'admin', 'mod', 'player'],
+    'admin': ['mod', 'player'],
+    'mod': ['player'],
+    'player': []
+}
 
 class UserCreate(BaseModel):
-    first_name: str
-    last_name: str
-    email: str
+    # Enter by admin
     discord_id: str | None = "inconnu"
     rp_first_name: str | None = "inconnu"
     rp_last_name: str | None = "inconnu"
-    rp_birthdate: date | None = date.today()
-    rp_gender: str | None = "inconnu"
     rp_grade: str | None = "inconnu"
     rp_affectation: str | None = "inconnu"
+    rp_qualif: str | None = "afp" 
     rp_nipol: str | None = "inconnu"
     rp_server: str | None = "inconnu"
     rp_service: str | None = "inconnu"  # Police(PN) / Gendarmerie(GN) / Police Municipale(PM)
+    rp_qualif: str | None = "afp"
+
     model_config = ConfigDict(from_attributes=True)
 
 class UserAdminView(BaseModel):
@@ -69,6 +74,10 @@ class UserAdminView(BaseModel):
 class SetUserAdminRequest(BaseModel):
     is_admin: bool
 
+class UserAfterCreation(BaseModel):
+    id: int
+    rp_nipol: int
+    temp_password: str
 
 def get_db():
     db = SessionLocal()
@@ -80,7 +89,15 @@ def get_db():
 db_dependency = Annotated[Session, Depends(get_db)]
 user_dependency = Annotated[models.Users, Depends(get_current_user)]
 
-protectedUsers = [""]
+protectedUsers = ["admin@admin.com"]
+
+def generate_temp_password():
+    import random
+    import string
+    length = 8
+    characters = string.ascii_letters + string.digits
+    temp_password = ''.join(random.choice(characters) for i in range(length))
+    return temp_password
 
 @router.get("/users/", response_model=List[UserAdminView])
 async def read_all_users(db: db_dependency, request: Request, current_user: user_dependency):
@@ -100,6 +117,8 @@ async def read_specific_user(user_id: int, db: db_dependency, request: Request, 
 @router.post("/set_user_privileges/{user_id}")
 async def set_user_privileges(user_id: int, db: db_dependency, request: Request, current_user: user_dependency, privilege: str = Body(..., embed=True)):
     list_of_privileges = ["player", "mod", "admin", "owner"]
+    if current_user.privileges != "owner": # type: ignore
+        raise HTTPException(status_code=403, detail="Only owner can change user privileges")
     if privilege not in list_of_privileges:
         raise HTTPException(status_code=400, detail=f"Invalid privilege. Must be one of: {', '.join(list_of_privileges)}")
     user = db.query(Users).filter(Users.id == user_id).first()
@@ -107,6 +126,7 @@ async def set_user_privileges(user_id: int, db: db_dependency, request: Request,
         raise HTTPException(status_code=404, detail="User not found")
     if user.email in protectedUsers:
         raise HTTPException(status_code=403, detail="Cannot change privileges of protected users")
+    
     user.privileges = privilege # type: ignore
     db.commit()
     api_log("admin.set_user_privileges", level="INFO", request=request, tags=["admin", "set_privileges"], user_id=current_user.id,email=current_user.email, data={"changed_user_id": user.id, "changed_user_nipol": user.rp_nipol, "new_privilege": privilege}, correlation_id=request.headers.get("x-correlation-id")) # type: ignore
@@ -120,42 +140,72 @@ async def delete_user(user_id: int, db: db_dependency, request: Request, current
         raise HTTPException(status_code=404, detail="User not found")
     if user.email in protectedUsers:
         raise HTTPException(status_code=403, detail="Cannot delete protected users")
+    if user.privileges not in dicoAllowToChange[current_user.privileges]: # type: ignore
+        raise HTTPException(status_code=403, detail="Cannot delete user with equal or higher privileges")
+    if user.id == current_user.id: # type: ignore
+        raise HTTPException(status_code=403, detail="Cannot delete yourself")
+    
+    notifications = db.query(models.Notifications).filter(models.Notifications.user_id == user.id).all()
+    for notification in notifications:
+        db.delete(notification)
+    db.commit()
     db.delete(user)
     db.commit()
     api_log("admin.delete_user", level="WARNING", request=request, tags=["admin", "delete_user"], user_id=current_user.id,email=current_user.email, data={"deleted_user_id": user.id, "deleted_user_nipol": user.rp_nipol}, correlation_id=request.headers.get("x-correlation-id")) # type: ignore
     return {"message": "User deleted successfully"}
 
 # ---------- Registration ----------
-@router.post("/register/", status_code=status.HTTP_201_CREATED)
+@router.post("/register/", status_code=status.HTTP_201_CREATED, response_model=UserAfterCreation)
 async def register_user(user: UserCreate, db: db_dependency, request: Request, current_user: user_dependency):
     # Hash the password
-    hashed_password = bcrypt_context.hash("temporaire")  # Default password, should be changed by user later
+    temp_password = generate_temp_password()
+    hashed_password = bcrypt_context.hash(temp_password)  # Default password, should be changed by user later
     db_user = Users(
-        first_name=user.first_name,
-        last_name=user.last_name,
-        email=user.email,
-        password=hashed_password,
-        discord_id=user.discord_id,
-        inscription_status="pending",
-        inscription_date=date.today(),
-        rp_gender=user.rp_gender,
+        # Entered by user later
+        first_name="inconnu",
+        last_name="inconnu",
+        email="inconnu",
+        rp_gender="male",
+        rp_birthdate=date.today(),
+
+        # Entered by admin
         rp_first_name=user.rp_first_name,
         rp_last_name=user.rp_last_name,
-        rp_birthdate=user.rp_birthdate,
+        discord_id=user.discord_id,
         rp_grade=user.rp_grade,
         rp_affectation=user.rp_affectation,
-        rp_qualif="afp", #TODO: DEFAUT A CHANGER (Voir IRL)
+        rp_qualif=user.rp_qualif,
         rp_nipol=user.rp_nipol,
         rp_server=user.rp_server,
         rp_service=user.rp_service,
-        privileges='player',
+
+        # Force or Generated
+        password=hashed_password,
+        inscription_status="pending",
+        inscription_date=date.today(),
+        privileges="player",
+        temp_password=True,  
+
     )
+
+
     try:
         db.add(db_user)
         db.commit()
         db.refresh(db_user)
+        new_user = db.query(Users).filter(Users.rp_nipol == db_user.rp_nipol).first()
+        if not new_user:
+            raise HTTPException(status_code=500, detail="User creation failed")
+        db_notifications = models.Notifications(
+            user_id=new_user.id,  # type: ignore
+            title="Complétez votre inscription",
+            message="Rendez-vous dans votre profil pour compléter votre inscription et choisir votre mot de passe. \n Ensuite, un administrateur validera votre inscription.",
+            redirect_to="/profile",
+        )
+        db.add(db_notifications)
+        db.commit()
         api_log("admin.register_user", level="INFO", request=request, tags=["admin", "register_user"], user_id=current_user.id,email=current_user.email,data={"created_user_id": db_user.id, "created_user_nipol": db_user.rp_nipol} ,correlation_id=request.headers.get("x-correlation-id")) # type: ignore
-        return {"message": "User registered successfully"}
+        return {"id": db_user.id, "rp_nipol": db_user.rp_nipol, "temp_password": temp_password}
     except Exception as e:
         db.rollback()
         if db.query(Users).filter(Users.rp_nipol == user.rp_nipol).first():
@@ -180,7 +230,6 @@ class UserUpdate(BaseModel):
     rp_nipol: str | None = None
     rp_server: str | None = None
     rp_service: str | None = None  # Police(PN) / Gendarmerie(GN) / Police Municipale(PM)
-    privileges: str | None = None
     model_config = ConfigDict(from_attributes=True)
 
 @router.patch("/users_update/{user_id}", response_model=UserAdminView)
@@ -190,6 +239,8 @@ async def update_user(user_id: int, user_update: UserUpdate, db: db_dependency, 
         raise HTTPException(status_code=404, detail="User not found")
     if user.email in protectedUsers:
         raise HTTPException(status_code=403, detail="Cannot update protected users")
+    if user.privileges not in dicoAllowToChange[current_user.privileges]: # type: ignore
+        raise HTTPException(status_code=403, detail="Cannot update user with equal or higher privileges")
 
     update_data = user_update.model_dump(exclude_unset=True)
 
@@ -207,24 +258,27 @@ async def update_user(user_id: int, user_update: UserUpdate, db: db_dependency, 
     api_log("admin.update_user", level="INFO", request=request, tags=["admin", "update_user"], user_id=current_user.id,email=current_user.email, data={"updated_user_id": user.id, "updated_user_nipol": user.rp_nipol}, correlation_id=request.headers.get("x-correlation-id")) # type: ignore
     return user
 
-class PasswordUpdate(BaseModel):
-    new_password: str
 
-@router.post("/users/{user_id}/password")
-async def update_password(user_id: int, body: PasswordUpdate, db: db_dependency, request: Request, current_user: user_dependency):
+@router.post("/users/{user_id}/password", response_model=UserAfterCreation)
+async def update_password(user_id: int, db: db_dependency, request: Request, current_user: user_dependency):
     user = db.query(Users).filter(Users.id == user_id).first()
     if not user:
         raise HTTPException(status_code=404, detail="User not found")
     if user.email in protectedUsers:
         raise HTTPException(status_code=403, detail="Cannot update protected users")
-    user.password = bcrypt_context.hash(body.new_password)  # type: ignore
+    if user.privileges not in dicoAllowToChange[current_user.privileges]: # type: ignore
+        raise HTTPException(status_code=403, detail="Cannot update user with equal or higher privileges")
+    temp_password = generate_temp_password()
+    user.password = bcrypt_context.hash(temp_password)  # type: ignore
     user.temp_password = True # type: ignore
     db.commit()
     api_log("admin.update_password", level="CRITICAL", request=request, tags=["admin", "update_password"], user_id=current_user.id,email=current_user.email,data={"updated_user_id": user.id, "updated_user_nipol": user.rp_nipol} ,correlation_id=request.headers.get("x-correlation-id")) # type: ignore
-    return {"message": "Password updated"}
+    return {"id": user.id, "rp_nipol": user.rp_nipol, "temp_password": temp_password} # type: ignore
 
 @router.post("/users/disconnect_all")
 async def disconnect_all_users(db: db_dependency, request: Request, current_user: user_dependency):
+    if current_user.privileges != "owner": # type: ignore
+        raise HTTPException(status_code=403, detail="Only owner can disconnect all users")
     users = db.query(Users).all()
     for user in users:
         user.token_version += 1  # type: ignore # Incrémente la version du token pour forcer la déconnexion
@@ -239,6 +293,8 @@ async def disconnect_user(user_id: int, db: db_dependency, request: Request, cur
         raise HTTPException(status_code=404, detail="User not found")
     if user.email in protectedUsers:
         raise HTTPException(status_code=403, detail="Cannot disconnect protected users")
+    if user.privileges not in dicoAllowToChange[current_user.privileges]: # type: ignore
+        raise HTTPException(status_code=403, detail="Cannot disconnect user with equal or higher privileges")
     user.token_version += 1  # type: ignore # Incrémente la version du token pour forcer la déconnexion
     db.commit()
     api_log("admin.disconnect_user", level="CRITICAL", request=request, tags=["admin", "disconnect_user"], user_id=current_user.id,email=current_user.email, data={"disconnected_user_id": user.id, "disconnected_user_nipol": user.rp_nipol}, correlation_id=request.headers.get("x-correlation-id")) # type: ignore
